move some fixtures
