+++ tests/unit/conftest.py
@@ -5,10 +5,13 @@ import pytest
 import json
 import sys
 
+from unittest import mock
 from shutil import copytree
 from artifactory import _ArtifactoryAccessor, _FakePathTemplate, ArtifactoryPath
 
 from galactory import create_app
+
+from galactory.utilities import discover_collections as original_discover_collections
 
 
 @pytest.fixture
@@ -143,3 +146,21 @@ def mock_artifactory_path(mock_artifactory_accessor, virtual_fs_repo):
 @pytest.fixture
 def repository(mock_artifactory_path):
     return mock_artifactory_path('http://artifactory.example.com/repo/subpath')
+
+
+@pytest.fixture
+def manifest_loader():
+    def _load(repo):
+        return repo._galactory_get_manifest()
+
+    loader = mock.Mock(wraps=_load)
+
+    with mock.patch('galactory.utilities.load_manifest_from_archive', loader):
+        yield loader
+
+
+@pytest.fixture
+def discover_collections(manifest_loader):
+    _discover = mock.Mock(wraps=original_discover_collections)
+    with mock.patch('galactory.utilities.discover_collections', _discover):
+        yield _discover
